fix(rpc): shut down gracefully after completed RPC call
CLI commands running over RPC were hanging after completion thanks to the RPC client side
instance never firing on the shutdown channel.

The fix for this was one line: add `go inst.waitForAllDone()` to lib.NewInstance when returning
an RPC-backed instance. All other changes are in service of testing that line.

* Moved the `ServeRPC` method down from the API onto the instance itself, making testing easier
* colocated all the rpc stuff into one file: rpc.go, which should make it easier to refactor
  in the future
* re-worked some of the NetworkIntegrationTestRunner in lib, then realized it was overkill
  for the RPC test.
