+++ api/api.go
@@ -5,42 +5,32 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"io/ioutil"
-	stdlog "log"
 	"net/http"
-	"net/rpc"
 	"time"
 
 	golog "github.com/ipfs/go-log"
-	ma "github.com/multiformats/go-multiaddr"
-	manet "github.com/multiformats/go-multiaddr-net"
 	"github.com/qri-io/apiutil"
 	"github.com/qri-io/qfs/cafs"
 	"github.com/qri-io/qri/lib"
 	"github.com/qri-io/qri/version"
 )
 
-var log = golog.Logger("qriapi")
-
-// APIVersion is the version string that is written in API responses
-var APIVersion = version.String
+var (
+	log = golog.Logger("qriapi")
+	// APIVersion is the version string that is written in API responses
+	APIVersion = version.String
+)
 
 // LocalHostIP is the IP address for localhost
-const LocalHostIP = "127.0.0.1"
-
-// DefaultTemplateHash is the hash of the default render template
-const DefaultTemplateHash = "/ipfs/QmeqeRTf2Cvkqdx4xUdWi1nJB2TgCyxmemsL3H4f1eTBaw"
-
-// TemplateUpdateAddress is the URI for the template update
-const TemplateUpdateAddress = "/ipns/defaulttmpl.qri.io"
+const (
+	LocalHostIP = "127.0.0.1"
+	// DefaultTemplateHash is the hash of the default render template
+	DefaultTemplateHash = "/ipfs/QmeqeRTf2Cvkqdx4xUdWi1nJB2TgCyxmemsL3H4f1eTBaw"
+	// TemplateUpdateAddress is the URI for the template update
+	TemplateUpdateAddress = "/ipns/defaulttmpl.qri.io"
+)
 
 func init() {
-	// We don't use the log package, and the net/rpc package spits out some complaints b/c
-	// a few methods don't conform to the proper signature (comment this out & run 'qri connect' to see errors)
-	// so we're disabling the log package for now. This is potentially very stupid.
-	// TODO (b5): remove dep on net/rpc package entirely
-	stdlog.SetOutput(ioutil.Discard)
-
 	golog.SetLogLevel("qriapi", "info")
 }
 
@@ -126,42 +116,6 @@ func (s Server) Serve(ctx context.Context) (err error) {
 
 	// http.ListenAndServe will not return unless there's an error
 	return StartServer(cfg.API, server)
-}
-
-// ServeRPC checks for a configured RPC port, and registers a listner if so
-func (s Server) ServeRPC(ctx context.Context) {
-	cfg := s.Config()
-	if !cfg.RPC.Enabled || cfg.RPC.Address == "" {
-		return
-	}
-	maAddress := cfg.RPC.Address
-	addr, err := ma.NewMultiaddr(maAddress)
-	if err != nil {
-		log.Errorf("cannot start RPC: error parsing RPC address %s: %w", maAddress, err.Error())
-	}
-
-	mal, err := manet.Listen(addr)
-	if err != nil {
-		log.Infof("RPC listen on address %d error: %w", cfg.RPC.Address, err)
-		return
-	}
-	listener := manet.NetListener(mal)
-
-	for _, rcvr := range lib.Receivers(s.Instance) {
-		if err := rpc.Register(rcvr); err != nil {
-			log.Errorf("cannot start RPC: error registering RPC receiver %s: %w", rcvr.CoreRequestsName(), err.Error())
-			return
-		}
-	}
-
-	go func() {
-		<-ctx.Done()
-		log.Info("closing RPC")
-		listener.Close()
-	}()
-
-	rpc.Accept(listener)
-	return
 }
 
 // HandleIPFSPath responds to IPFS Hash requests with raw data
